Lift the definite-assignment analysis out of CSharpPrinter (printer decomposition 3/3)
Final slice of the CSharpPrinter decomposition. Moves the
definite-assignment analysis behind the printer's `= default` elision —
the ~430-LOC DefiniteFlow walk and its CFG fixpoint (ComputeReadBeforeAssign
plus the DefiniteFlow enum) — into a new self-contained analyzer,
DefiniteAssignment.Compute(function, labelTargets, facts). The printer now
consumes the fact (the read-before-assign set) instead of computing it
inline.

This is the analysis whose result decides which locals keep their
zero-initializer (a bare declaration of a read-before-assign local is
CS0165) versus declaring bare (the initializer is a dead store the IL
never carried — locals lean on .locals init). It is the same definite
assignment the C# compiler runs.

Extracted as an analyzer the printer calls, NOT an IIrPass: the pass
contract is to rewrite the tree and "never side-channel state", whereas
this is fact-producing, not tree-rewriting. Forcing it into a pass would
require exactly the side-channel IR state the contract forbids, so an
analyzer class is the honest tool.

The walk touched only two printer fields (_labelTargets, _facts), now
explicit parameters; the lone naming concern (facts.LocalNames) stays in
the printer, which owns LocalName. Verbatim move — zero behavior change.
Printer 1967 -> 1530 LOC (2203 -> 1530 across all three slices).
348/348 decompiler tests green (incl. DataflowFactsTests).
